placeholder button and fix for sidebar
there's now a placeholder button for the casting action in the main app
page.
I fixed an issue with the sidebar. The right sidebar was still
displaying to the right of the main content, and could be scrolled to
(horizontally). I changed its "position" css class to be "fixed" instead
of "absolute".