cast button now disables when the user selects bad options
Since the cast button is now rendered conditionally with different
tailwind classes, it was starting to clutter the App component, so it
was moved out to its own svelte component.

There is a new alert component to help clarify this.

Sidebars were given a tailwind "z-50" so that they appear over anything
else, even when the error is rendered, so that it does not get in a
user's way while they try to fix the error.

The "writeSpellParameters" store was altered to be a default writable
store, since its custom functionality proved to be unnecessary. The
tests that covered that custom functionality have thus been removed
also.

Lastly, a new eslint rule was added so that the compiler/ intellisense
does not complain about "test" and other jest global variables.