new animations for the results list and error message!
There was some extra work necessary to add animations properly, since
svelte requiers a unique identifier for each item rendered in an {#each}
block so that it can operate on them effectively/ correctly. Basically,
the running list of casting results was turned into a list of
objects in the shape { creatures, id }.

Additionally, the animations that I settled on for the results list
became quite confusing to me. I understand what's going on at a high
level, but the specifics of the logic are beyond me at the moment. The
code was extracted right from the svelte tutorial/ documentation, and
does not affect the operation of the app from a DOM perspective, and so
I am not worried at the moment that I don't quite get it. It can (and
should!) be fiddled with later.