Bug fix for background scrolling while in menu
Programmatically add the "h-full" and "overflow-hidden" classes to a new
div that contains all of the content that will be in the "body" of the
html, so that whenver either of the sidebar menus are opened all of the
"body" content is locked in place and limited to the size of the screen
until the menu is closed.

This solution has the benefit over using "position-fixed" that the
scroll history is not reset when the sidebar menu is closed.