Add feature: CR filtered by other options
When the user selects options like the spell that they are casting, the
source books that they are using, or the terrains that they want to
summon creatures for, the challenge rating drop-down select will only
show the challenge ratings that would summon a non-empty list of
creatures.

This may prove to be performance-degrading on devices with less
computational power, and will have to be closely monitored with my
testers.

Additionally, the tests for the spell parameters selection component
have been refactored to use the "describe..it..." format, instead of
just using the "test" function, per the recommendation in the thoughtbot
guides.