Add an "About" component, modify the sidebar
There is a new component that lists some information about the
application. To display it in the desktop application, it seemed best to
have it slide in from a sidebar since it contains so much text (when the
details tags are opened).

The current Sidebar component did not support anything except full
width, which looks a bit off on a desktop display. The component was
modified to support a "halfScreen" prop so that it can render out to
only half-width.

With those changes made, since the user could still see half of the
application, it seems odd that they can interact with the app WITHOUT
closing the sidebar. Thus, a new on-click handler was added to the div
that contains the body of the app so that if the user clicks anywhere in
there, the sidebar will be closed.