changes to the mobile navigation
Added the About component as the child of the right sidebar on mobile,
much like it is on desktop. In doing so, I moved the SelectSourcebooks
component into the left-hand sidebar, which has the added benefit that
users are less likely to miss the sourcebook menu, as some of my users
have told me happened to them.

Made major edits to the tests for the ResponsiveApp component, since
many of them were interacting with the right-hand mobile hamburger menu
and expecting a sourcebook selection menu.