About the app (#18)
There is a new "About" component that lists some information about the
application. To display it in the application, it seemed best to
have it slide in from a sidebar since it contains so much text (when the
details tags are opened).

The current Sidebar component did not support anything except full
width, which looks a bit off on a desktop display. The component was
modified to support a "halfScreen" prop so that it can render out to
only half-width.

With those changes made, since the user could still see half of the
application, it seems odd that they can interact with the app WITHOUT
closing the sidebar. Thus, a new on-click handler was added to the div
that contains the body of the app (for screens wider than 1024 px) so that if the user clicks anywhere in
there, the sidebar will be closed.

Minor style changes, including a clearer border for Sidebars and some hover classes for buttons and the ResultBox.

Added the About component as the child of the right sidebar on mobile,
much like it is on desktop. In doing so, I moved the SelectSourcebooks
component into the left-hand sidebar, which has the added benefit that
users are less likely to miss the sourcebook menu, as some of my users
have told me happened to them.

Made major edits to the tests for the ResponsiveApp component, since
many of them were interacting with the right-hand mobile hamburger menu
and expecting a sourcebook selection menu.